requirements.txt
absl-py==2.1.0
apex==0.1
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
filelock==3.13.1
flatbuffers==25.2.10
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
googledrivedownloader==1.1.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
jinja2==3.1.4
joblib==1.4.2
keras==3.9.0
libclang==18.1.1
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.5.82
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.5.82
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.5.82
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.3.61
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.6.82
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.3.83
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.1.3
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.23.4
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu11==11.8.86
opt-einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.0.0
protobuf==5.29.3
pygments==2.19.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-logger==0.1.0
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.6.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
triton==3.1.0
typing-extensions==4.12.2
urllib3==2.3.0
werkzeug==3.1.3
wrapt==1.17.2
